Backend/requirements.txt
fastapi==0.136.1
uvicorn==0.46.0
SQLAlchemy==2.0.49
psycopg2-binary==2.9.12
python-jose==3.5.0
passlib==1.7.4
bcrypt==4.0.1
cryptography==48.0.0
pydantic==2.13.4
pydantic-settings==2.14.1
python-dotenv==1.2.2
email-validator==2.3.0
pandas==3.0.2
openpyxl==3.1.5
pillow==12.2.0
python-multipart==0.0.28